add EditTaskViewModel, layout to edit task and logic to EditTaskFragment

+++ Tasks/app/src/main/java/com/example/tasks/TaskDao.kt
@@ -22,7 +22,7 @@ interface TaskDao {
     suspend fun delete(task: Task)
 
     @Query("SELECT * FROM task_table WHERE taskId = :taskId")
-    fun get(taskId: Int): LiveData<Task>
+    fun get(taskId: Long): LiveData<Task>
 
     @Query("SELECT * FROM task_table ORDER BY taskId DESC")
     fun getAll(): LiveData<List<Task>>
